[libpas] Use futex-based locking in Windows / FreeBSD https://bugs.webkit.org/show_bug.cgi?id=306203 rdar://168860161
Reviewed by Justin Michaud and Sosuke Suzuki.

This patch adds futex locking in Windows / FreeBSD as well.
Also remove spinlock, as all libpas-suppoting platforms support
this new locking. Also move slow path code to C side to avoid bloating
code for locking.

* Source/WTF/wtf/PlatformJSCOnly.cmake:
* Source/WTF/wtf/PlatformWin.cmake:
* Source/bmalloc/libpas/src/libpas/pas_config.h:
* Source/bmalloc/libpas/src/libpas/pas_lock.c:
(pas_lock_futex_wait):
(pas_lock_futex_wake):
(pas_lock_lock_slow):
(pas_lock_unlock_slow):
* Source/bmalloc/libpas/src/libpas/pas_lock.h:
(pas_lock_construct_disabled):
(pas_lock_mutex_setname):
(pas_lock_lock):
(pas_lock_try_lock):
(pas_lock_unlock):
(pas_lock_test_held):
(pas_lock_assert_held):
(pas_lock_testing_assert_held):
(pas_lock_futex_wait):
(pas_lock_futex_wake):

Canonical link: https://commits.webkit.org/306197@main
