+++ Source/bmalloc/libpas/src/libpas/pas_config.h
@@ -138,11 +138,5 @@
 
 #define PAS_ALLOCATOR_INDEX_BYTES        4
 
-#if PAS_OS(DARWIN) || PAS_OS(LINUX) || PAS_PLATFORM(PLAYSTATION)
-#define PAS_USE_SPINLOCKS                0
-#else
-#define PAS_USE_SPINLOCKS                1
-#endif
-
 #endif /* PAS_CONFIG_H */
